refactor(ai-engine): Split reranking_engine.py into reranking/ subpackage
Split 55K, 1492-line reranking_engine.py into modular package per Issue #1140:

- Created ai-engine/search/reranking/ package with:
  - __init__.py: Public API re-exports
  - base.py: ReRankingStrategy, ReRankingFeature, ReRankingResult dataclasses
  - feature_reranker.py: FeatureBasedReRanker (670 lines)
  - contextual_reranker.py: ContextualReRanker (125 lines)
  - ensemble_reranker.py: EnsembleReRanker (115 lines)
  - cross_encoder.py: CrossEncoderReRanker (197 lines)
  - neural_reranker.py: NeuralReRanker (145 lines)
  - hybrid_reranker.py: HybridReRanker (146 lines)

- Legacy reranking_engine.py preserved with DeprecationWarning for backwards compatibility
- Updated pyproject.toml with search/reranking/ ignores (F401, C901)
- All 73 search tests pass, 17 reranking-specific tests pass
