+++ ai-engine/search/reranking/__init__.py
@@ -1,0 +1,26 @@
+"""
+Re-ranking module for improving search result quality.
+
+Provides various re-ranking strategies to improve the final ordering
+of search results based on additional relevance signals.
+"""
+
+from .base import ReRankingFeature, ReRankingResult, ReRankingStrategy
+from .feature_reranker import FeatureBasedReRanker
+from .contextual_reranker import ContextualReRanker
+from .ensemble_reranker import EnsembleReRanker
+from .cross_encoder import CrossEncoderReRanker
+from .neural_reranker import NeuralReRanker
+from .hybrid_reranker import HybridReRanker
+
+__all__ = [
+    "ReRankingStrategy",
+    "ReRankingFeature",
+    "ReRankingResult",
+    "FeatureBasedReRanker",
+    "ContextualReRanker",
+    "EnsembleReRanker",
+    "CrossEncoderReRanker",
+    "NeuralReRanker",
+    "HybridReRanker",
+]
